+++ packages/wallet-sdk/.eslintrc.js
@@ -195,6 +195,7 @@ module.exports = {
       },
     ],
     "simple-import-sort/exports": "error",
+    "no-restricted-imports": ["error", "eth-rpc-errors"],
     "space-before-function-paren": "off",
     "space-in-parens": ["off", "never"],
     "spaced-comment": [
